feat(struct): create schemas file
